Seeing if I can fix the retry issue by backing up the r_MAC
I save a copy when it is about to be overwritten, and then use it if I get a dup.

I am guessing I dont need to worry about updates to the SC if I dont decode because the MAC will not be updated. It is the CP's issue. Yay.

Interesting. Which means it is likely the Inception that is not decoding the serial occasionally. Oops.